requirements.txt

future
futures
# fusepy==3.0.1

# Neo4j
# gremlinpython==3.2.6

# OrientDB
# gremlinpython==3.3.0
gremlinpython==3.3.7

pika==1.1.0
kafka-python

simplejson
# 0.1.7/0.1.8 has bugs
flatten_json==0.1.6
addict

PyYAML
pystache
# flask
Flask==1.1.1
flask-blueprint
Flask-RESTful
flask-cors
flask-swagger
flask-socketio
Flask-Sockets==0.2.1
gevent==1.4.0
gevent-websocket==0.10.1

greenlet==0.4.16

importlib-metadata
importlib-resources
jsonschema

# flask_graphql
Flask-GraphQL==2.0.0
# graphene
graphene==2.1.8

graphql-core==2.2.1
graphql-relay==2.0.0
graphql-server-core==1.1.1
# graphql-ws==0.3.0
graphql-ws==0.3.1

Werkzeug==0.15.5

# Rx==3.1.1
Rx==1.6.1

# 
# REST/GraphQL test client
# 
requests==2.26.0
python-graphql-client==0.4.3
